@@ -36,36 +36,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "7.3.7",
-    "@angular/animations": "7.2.11",
-    "@angular/cdk": "7.3.6",
-    "@angular/common": "7.2.11",
-    "@angular/compiler": "7.2.11",
-    "@angular/core": "7.2.11",
-    "@angular/flex-layout": "7.0.0-beta.19",
-    "@angular/forms": "7.2.11",
-    "@angular/http": "7.2.11",
-    "@angular/material": "7.3.6",
-    "@angular/platform-browser": "7.2.11",
-    "@angular/platform-browser-dynamic": "7.2.11",
-    "@angular/router": "7.2.11",
+    "@angular/animations": "8.0.0",
+    "@angular/cdk": "8.0.0",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/flex-layout": "8.0.0-beta.26",
+    "@angular/forms": "8.0.0",
+    "@angular/http": "8.0.0-beta.10",
+    "@angular/material": "8.0.0",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
     "@ngx-translate/core": "11.0.1",
     "@ngx-translate/http-loader": "4.0.0",
-    "@swimlane/ngx-datatable": "14.0.0",
-    "@types/d3": "4.12.0",
-    "@types/d3-dsv": "1.0.32",
-    "@types/pixi.js": "4.8.7",
+    "@swimlane/ngx-datatable": "15.0.1",
     "angular-calendar": "0.26.4",
     "angular-draggable-droppable": "4.0.2",
     "angular-resizable-element": "3.2.2",
-    "angular-svg-icon": "7.0.1",
-    "angular-tree-component": "8.3.0",
+    "angular-svg-icon": "7.2.1",
+    "angular-tree-component": "8.4.0",
     "angular2-uuid": "1.1.1",
     "chart.js": "2.7.3",
     "chartist": "0.11.0",
     "chartist-plugin-legend": "0.6.2",
     "classlist.js": "1.1.20150312",
-    "commander": "2.19.0",
     "core-js": "2.6.0",
     "cron-parser": "2.7.3",
     "d3": "4.12.2",
@@ -98,45 +93,50 @@
     "pixi.js": "4.8.7",
     "popmotion": "8.3.3",
     "primeicons": "1.0.0",
-    "primeng": "6.1.7",
+    "primeng": "8.0.0-rc.1",
     "quill": "1.3.6",
-    "raven-js": "3.27.0",
+    "raven-js": "3.27.1",
     "rxjs": "6.3.3",
     "rxjs-compat": "6.3.3",
     "stream": "0.0.2",
     "stylefire": "2.2.0",
-    "webpack-cli": "^3.2.1",
     "what-is-my-ip-address": "1.0.3",
     "xterm": "2.9.2",
-    "zone.js": "0.8.26"
+    "zone.js": "0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.13.7",
-    "@angular-devkit/build-optimizer": "0.13.7",
-    "@angular/cli": "7.3.7",
-    "@angular/compiler-cli": "7.2.11",
+    "@angular-devkit/core": "8.0.2",
+    "@angular-devkit/build-angular": "0.800.2",
+    "@angular-devkit/build-optimizer": "0.800.2",
+    "@angular/cli": "8.0.2",
+    "@angular/compiler-cli": "8.0.0",
+    "webpack-cli": "3.3.3",
+    "@types/d3": "4.12.0",
+    "@types/d3-dsv": "1.0.32",
+    "@types/pixi.js": "4.8.7",
+    "commander": "2.20.0",
     "@biesbjerg/ngx-translate-extract": "2.3.4",
     "@types/c3": "0.4.52",
     "@types/chartist": "0.9.43",
     "@types/hopscotch": "0.2.30",
     "@types/jasmine": "2.8.11",
-    "@types/lodash": "4.14.118",
+    "@types/lodash": "4.14.134",
     "@types/node": "10.12.7",
     "c3": "0.4.18",
-    "codelyzer": "4.5.0",
+    "codelyzer": "5.1.0",
     "enhanced-resolve": "4.1.0",
     "jasmine-core": "3.3.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "3.1.1",
+    "karma": "4.1.0",
     "karma-chrome-launcher": "2.2.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "2.0.4",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "~1.4.0",
+    "karma-cli": "2.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.5",
+    "karma-jasmine": "2.0.1",
+    "karma-jasmine-html-reporter": "1.4.2",
     "po2json": "1.0.0-alpha",
-    "protractor": "5.4.1",
-    "ts-node": "7.0.1",
-    "tslint": "5.11.0",
-    "typescript": "3.1.6"
+    "protractor": "5.4.2",
+    "ts-node": "8.2.0",
+    "tslint": "5.17.0",
+    "typescript": "3.4.3"
   }
 }
