@@ -6,7 +6,6 @@ import {
   ViewChild,
   ViewChildren,
   AfterViewInit,
-  Query,
 } from '@angular/core';
 import * as _ from 'lodash';
 import { Router, ActivatedRoute } from '@angular/router';
@@ -49,10 +48,10 @@ export class ManagerComponent implements OnInit, OnDestroy, AfterViewInit {
   public original_disks: Array < any >;
   public orig_suggestable_disks: Array < any >;
   public error: string;
-  @ViewChild('disksdnd') disksdnd;
+  @ViewChild('disksdnd', { static: true}) disksdnd;
   @ViewChildren(VdevComponent) vdevComponents: QueryList < VdevComponent > ;
   @ViewChildren(DiskComponent) diskComponents: QueryList < DiskComponent > ;
-  @ViewChild(DatatableComponent) table: DatatableComponent;
+  @ViewChild(DatatableComponent, { static: false}) table: DatatableComponent;
   public temp = [];
 
   public name: string;
@@ -70,6 +69,7 @@ export class ManagerComponent implements OnInit, OnDestroy, AfterViewInit {
   protected existing_pools = [];
   public poolError = null;
   public isFooterConsoleOpen: boolean;
+  public loaderOpen = false;
 
   public submitTitle = T("Create");
   protected extendedSubmitTitle = T("Extend");
@@ -185,7 +185,7 @@ export class ManagerComponent implements OnInit, OnDestroy, AfterViewInit {
         }
       },
       (err) => {
-        new EntityUtils().handleError(this, err);
+        new EntityUtils().handleWSError(this, err, this.dialog);
       }
     );
     this.rest.get(this.resource_name + this.pk, {}).subscribe((res) => {
@@ -226,11 +226,19 @@ export class ManagerComponent implements OnInit, OnDestroy, AfterViewInit {
     }
     this.nameFilter = new RegExp('');
     this.capacityFilter = new RegExp('');
-    this.ws.call("notifier.get_disks", [true]).subscribe((res) => {
+
+  }
+
+  ngAfterViewInit() {
+    this.loader.open();
+    this.loaderOpen = true;
+    this.ws.call("disk.get_unused", [true]).subscribe((res) => {
+      this.loader.close();
+      this.loaderOpen = false;
       this.disks = [];
       for (let i in res) {
-        res[i]['real_capacity'] = res[i]['capacity'];
-        res[i]['capacity'] = (<any>window).filesize(res[i]['capacity'], {standard : "iec"});
+        res[i]['real_capacity'] = res[i]['size'];
+        res[i]['capacity'] = (<any>window).filesize(res[i]['size'], {standard : "iec"});
         this.disks.push(res[i]);
       }
 
@@ -254,20 +262,25 @@ export class ManagerComponent implements OnInit, OnDestroy, AfterViewInit {
       this. can_suggest = this.suggestable_disks.length < 11;
 
       this.temp = [...this.disks];
+    }, (err) => {
+      this.loader.close();
+      new EntityUtils().handleWSError(this, err, this.dialog)
     });
     this.ws.call('system.advanced.config').subscribe((res)=> {
       if (res) {
         this.isFooterConsoleOpen = res.consolemsg;
       }
     });
-  }
 
-  ngAfterViewInit() {
     if (!this.isNew) {
       setTimeout(() => { // goofy workaround for stupid angular error
         this.dialog.confirm(T("Warning"), helptext.manager_extend_warning, 
             false, T("Continue")).subscribe((res) => {
           if (!res) {
+            if (this.loaderOpen) {
+              this.loader.close();
+              this.loaderOpen = false;
+            }
             this.goBack();
           }
         });
