@@ -27,6 +27,7 @@ export interface Disk {
   name: string;
   smart_enabled: boolean;
   size: number;
+  model: string;
   description?: string;
   enclosure_slot?: any;
   expiretime?: any;
@@ -92,8 +93,11 @@ export class WidgetPoolComponent extends WidgetComponent implements AfterViewIni
   public currentDiskSet:number = 0;
   private simulateDiskArray:number;
   public displayValue: any;
+  public displayValueTotal: string;
   public diskSize: any;
   public diskSizeLabel: string;
+  public diskSizeTotal: any;
+  public diskSizeTotalLabel: string;
   @Input() configurable:boolean;
 
   constructor(public router: Router, public translate: TranslateService, public storage: StorageService){
@@ -175,12 +179,13 @@ export class WidgetPoolComponent extends WidgetComponent implements AfterViewIni
   }
 
   setDisksData(evt:CoreEvent){
-    for(let i in evt.data){
-      let disk:Disk = {
+    for(const i in evt.data){
+      const disk:Disk = {
         name:evt.data[i].name,
         smart_enabled:evt.data[i].togglesmart,
         size:Number(evt.data[i].size),
         description: evt.data[i].description,
+        model: evt.data[i].model,
         enclosure_slot: evt.data[i].enclosure_slot,
         expiretime: evt.data[i].expiretime,
         hddstandby: evt.data[i].hddstandby,
@@ -246,6 +251,14 @@ export class WidgetPoolComponent extends WidgetComponent implements AfterViewIni
     } else {
       this.diskSizeLabel = this.displayValue.slice(-3);
       this.diskSize = new Intl.NumberFormat().format(parseFloat(this.displayValue.slice(0, -4)))
+    }    
+    this.displayValueTotal = (<any>window).filesize(this.volumeData.avail + this.volumeData.used, {standard: "iec"});
+    if (this.displayValueTotal.slice(-2) === ' B') {
+      this.diskSizeTotalLabel = this.displayValueTotal.slice(-1);
+      this.diskSizeTotal = new Intl.NumberFormat().format(parseFloat(this.displayValueTotal.slice(0, -2)))
+    } else {
+      this.diskSizeTotalLabel = this.displayValueTotal.slice(-3);
+      this.diskSizeTotal = new Intl.NumberFormat().format(parseFloat(this.displayValueTotal.slice(0, -4)))
     }
     // Adds a zero to numbers with one (and only one) digit after the decimal
     if (this.diskSize.charAt(this.diskSize.length - 2) === '.' || this.diskSize.charAt(this.diskSize.length - 2) === ',') {
