@@ -3,6 +3,7 @@ import {Validators} from '@angular/forms';
 import {
   regexValidator
 } from '../../pages/common/entity/entity-form/validators/regex-validation';
+import { UserService } from 'app/services';
 
 export default {
 
@@ -16,7 +17,7 @@ bsdgrp_gid_validation: [ Validators.required, regexValidator(/^\d+$/) ],
 
 bsdgrp_group_placeholder: T('Name'),
 bsdgrp_group_tooltip: T('Enter an alphanumeric name for the group.'),
-bsdgrp_group_validation: [ Validators.required, regexValidator(/^[\w\_][\w\.-\_]*$/) ],
+bsdgrp_group_validation: [ Validators.required, Validators.pattern(UserService.VALIDATOR_NAME) ],
 
 bsdgrp_sudo_placeholder: T('Permit Sudo'),
 bsdgrp_sudo_tooltip: T('Allow group members to use <a\
