@@ -4,13 +4,13 @@ import { ActivatedRoute } from '@angular/router';
 
 import { FieldConfig } from 'app/pages/common/entity/entity-form/models/field-config.interface';
 import helptext from '../../../../helptext/task-calendar/replication/replication';
-import { WebSocketService, TaskService, KeychainCredentialService } from 'app/services';
+import { WebSocketService, TaskService, KeychainCredentialService, ReplicationService } from 'app/services';
 import * as _ from 'lodash';
 
 @Component({
     selector: 'app-replication-list',
     template: `<entity-form [conf]='this'></entity-form>`,
-    providers: [TaskService, KeychainCredentialService]
+    providers: [TaskService, KeychainCredentialService, ReplicationService]
 })
 export class ReplicationFormComponent {
 
@@ -179,10 +179,21 @@ export class ReplicationFormComponent {
                 }]
             }],
         }, {
-            type: 'input',
+            type: 'explorer',
             name: 'target_dataset_PUSH',
             placeholder: helptext.target_dataset_placeholder,
             tooltip: helptext.target_dataset_tooltip,
+            initial: '',
+            explorerType: 'directory',
+            customTemplateStringOptions: {
+                displayField: 'Path',
+                isExpandedField: 'expanded',
+                idField: 'uuid',
+                getChildren: this.getChildren.bind(this),
+                nodeHeight: 23,
+                allowDrag: false,
+                useVirtualScroll: false,
+            },
             required: true,
             validation: [Validators.required],
             isHidden: true,
@@ -194,10 +205,21 @@ export class ReplicationFormComponent {
                 }]
             }],
         }, {
-            type: 'input',
+            type: 'explorer',
             name: 'source_datasets_PULL',
             placeholder: helptext.source_datasets_placeholder,
             tooltip: helptext.source_datasets_placeholder,
+            initial: '',
+            explorerType: 'directory',
+            customTemplateStringOptions: {
+                displayField: 'Path',
+                isExpandedField: 'expanded',
+                idField: 'uuid',
+                getChildren: this.getChildren.bind(this),
+                nodeHeight: 23,
+                allowDrag: false,
+                useVirtualScroll: false,
+            },
             required: true,
             validation: [Validators.required],
             isHidden: true,
@@ -679,8 +701,12 @@ export class ReplicationFormComponent {
         },
     ]
 
-    constructor(private ws: WebSocketService, protected taskService: TaskService, private aroute: ActivatedRoute,
-        private keychainCredentialService: KeychainCredentialService) {
+    constructor(
+        private ws: WebSocketService,
+        protected taskService: TaskService,
+        private aroute: ActivatedRoute,
+        private keychainCredentialService: KeychainCredentialService,
+        private replicationService: ReplicationService) {
         const sshCredentialsField = _.find(this.fieldConfig, { name: 'ssh_credentials' });
         this.keychainCredentialService.getSSHConnections().subscribe(
             (res) => {
@@ -743,6 +769,19 @@ export class ReplicationFormComponent {
                 entityForm.setDisabled('schedule_end', toDisable, toDisable);
             }
         })
+
+        entityForm.formGroup.controls['ssh_credentials'].valueChanges.subscribe(
+            (res) => {
+                for (const item of ['target_dataset_PUSH', 'source_datasets_PULL']) {
+                    const explorerComponent = _.find(this.fieldConfig, {name: item}).customTemplateStringOptions.explorerComponent;
+                    explorerComponent.nodes = [{
+                        mountpoint: explorerComponent.config.initial,
+                        name: explorerComponent.config.initial,
+                        hasChildren: true
+                    }];
+                }
+            }
+        )
     }
 
     resourceTransformIncomingRestData(wsResponse) {
@@ -880,4 +919,23 @@ export class ReplicationFormComponent {
         }
     }
 
+    getChildren(node) {
+        for (const item of ['target_dataset_PUSH', 'source_datasets_PULL']) {
+            _.find(this.fieldConfig, {name: 'target_dataset_PUSH'}).hasErrors = false;
+        }
+
+        const transport = this.entityForm.formGroup.controls['transport'].value;
+        const sshCredentials = this.entityForm.formGroup.controls['ssh_credentials'].value;
+        if (sshCredentials == undefined || sshCredentials == '') {
+            for (const item of ['target_dataset_PUSH', 'source_datasets_PULL']) {
+                _.find(this.fieldConfig, {name: item}).hasErrors = true;
+                _.find(this.fieldConfig, {name: item}).errors = 'Please select a valid SSH Connection';
+            }
+            return;
+        }
+
+        return new Promise((resolve, reject) => {
+            resolve(this.replicationService.getRemoteDataset(transport,sshCredentials, this));
+        });
+    }
 }
